refactor(#147): 경로 및 명칭 수정
