+++ exbackend/src/domain/server/controllers/server-role-permission.controller.ts
@@ -13,7 +13,8 @@ import {
   ApiResponse,
   ApiBearerAuth,
 } from '@nestjs/swagger';
-import { JwtAuthGuard, CurrentUser, User } from '../../auth';
+import { JwtAuthGuard } from '../../auth';
+import { CurrentUser, User } from '../../user'
 import { ServerRolePermissionService } from '../services/server-role-permission.service';
 import {
   UpdateRolePermissionDto,
